requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
Deprecated==1.2.14
exceptiongroup==1.1.3
fastapi==0.103.1
FastAPI-SQLAlchemy==0.2.1
firebase-admin==6.2.0
gcloud==0.18.3
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
jwcrypto==1.5.0
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.5
oauth2client==4.1.3
proto-plus==1.22.3
protobuf==4.24.3
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dotenv==1.0.0
python-jwt==4.0.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
wrapt==1.15.0
